[FEAT/#9] 갤러리 이미지 가져오기 기능 구현
