+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 
     <application
         android:allowBackup="true"
@@ -27,7 +28,7 @@
         android:exported="false" />
         <activity
             android:name=".presentation.MainActivity"
-            android:exported="false" />
+            android:exported="true" />
     </application>
 
 </manifest>
