+++ app/src/main/java/org/sopt/sample/presentation/user/GalleryFragment.kt
@@ -1,0 +1,55 @@
+package org.sopt.sample.presentation.user
+
+import android.Manifest
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.fragment.app.Fragment
+import coil.load
+import org.sopt.sample.databinding.FragmentGalleryBinding
+
+class GalleryFragment : Fragment() {
+    private var _binding: FragmentGalleryBinding? = null
+    private val binding: FragmentGalleryBinding
+        get() = requireNotNull(_binding) { "binding value was null." }
+
+    private val launcher =
+        registerForActivityResult(ActivityResultContracts.PickMultipleVisualMedia()) {
+            binding.imgSample.load(it.get(0))
+            binding.imgSample2.load(it.get(1))
+        }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val permissionLauncher =
+            registerForActivityResult(ActivityResultContracts.RequestPermission()) {}
+        permissionLauncher.launch(Manifest.permission.READ_EXTERNAL_STORAGE)
+
+        binding.btnImage.setOnClickListener {
+            launcher.launch(PickVisualMediaRequest())
+        }
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    companion object {
+        fun newInstance(): GalleryFragment {
+            return GalleryFragment()
+        }
+    }
+}
