+++ app/src/main/res/layout/fragment_gallery.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/layout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".presentation.user.GalleryFragment">
+
+    <ImageView
+        android:id="@+id/img_sample"
+        android:layout_width="120dp"
+        android:layout_height="0dp"
+        android:layout_marginTop="120dp"
+        android:src="@drawable/ic_launcher_foreground"
+        app:layout_constraintDimensionRatio="1.0"
+        app:layout_constraintEnd_toStartOf="@id/img_sample2"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageView
+        android:id="@+id/img_sample2"
+        android:layout_width="120dp"
+        android:layout_height="0dp"
+        android:layout_marginTop="120dp"
+        android:src="@drawable/ic_launcher_foreground"
+        app:layout_constraintDimensionRatio="1.0"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toEndOf="@id/img_sample"
+        app:layout_constraintEnd_toEndOf="parent" />
+
+    <Button
+        android:id="@+id/btn_image"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="32dp"
+        android:padding="12dp"
+        android:text="이미지 가져오기"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/img_sample" />
+</androidx.constraintlayout.widget.ConstraintLayout>
