+++ app/src/main/res/layout/item_home_follower.xml
@@ -9,7 +9,7 @@
     android:paddingVertical="10dp">
 
     <ImageView
-        android:id="@+id/img_profile"
+        android:id="@+id/img_sample"
         android:layout_width="95dp"
         android:layout_height="0dp"
         app:layout_constraintDimensionRatio="1"
@@ -22,7 +22,7 @@
         android:id="@+id/txt_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/img_profile"
+        app:layout_constraintTop_toBottomOf="@id/img_sample"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         android:layout_marginTop="10dp"
