+++ app/src/main/res/values/strings.xml
@@ -11,7 +11,7 @@
     <string name="signup">SIGN UP</string>
     <string name="signup_complete">회원가입 완료</string>
     <string name="home">홈</string>
-    <string name="my_info">내 정보</string>
+    <string name="gallery">갤러리</string>
     <string name="setting">설정</string>
 
     <string name="msg_incorrect">잘못된 입력입니다.</string>
